requirements.txt
autopep8==2.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oracledb==1.2.1
pycodestyle==2.10.0
pycparser==2.21
python-dotenv==0.21.0
tomli==2.0.1
visitor==0.1.3
Werkzeug==2.2.2